Port Respawn System (#998)
This ports https://github.com/WWhiteDreamProject/wwdpublic/pull/56
The new Respawn System is one that allows players to return themselves
to the lobby after a configurable delay, while also requiring that they
respawn as a different character upon returning.

- [x] Finish the usual cleanup

:cl:
- add: Ported a Respawn System. This system allows players to return
themselves to the lobby, while also requiring that if they re-enter the
round, that they must do so on a different character.

---------

Co-authored-by: Spatison <137375981+Spatison@users.noreply.github.com>